chore(rpc-alt): upgrade jsonrpsee
## Description

Upgrade jsonrpsee we're using for this indexer to the latest version
(0.24.7) rather than the fork of 0.18.2 that we were using before. This
is because 0.24.7 has better support for middleware that can access the
parsed JSONRPC method (useful for measuring the time spent processing
each method).

## Test plan

Run the service and query it:

```
sui$ cargo run -p sui-indexer-alt-jsonrpc
```

```
curl -LX POST  "http://localhost:6000" \
    --header 'Content-Type: application/json' \
    --data '{
  "jsonrpc": "2.0",
  "id": 1,
  "method": "suix_getReferenceGasPrice",
  "params": []
}' | jq .
{
  "jsonrpc": "2.0",
  "id": 1,
  "result": "1000"
}
```